+++ app/webpacker/components/Panel/pages/CreateNewcomersPage/CompetitionsInput.jsx
